BGDIINF_SB-1525 validators in separate file
* moved validate_json_payload() in validators_serializer.py
* added logger.error() to validate_json_payload()
* linting of unrelated migrations file
