+++ app/stac_api/validators.py
@@ -187,27 +187,3 @@ def validate_item_properties_datetimes(
             properties_start_datetime,
             properties_end_datetime,
         )
-
-
-def validate_json_payload(serializer):
-    '''
-    Validate JSON payload and raise error, if extra payload or read-only fields
-    in payload are found
-    Args:
-        serializer: serializer for which payload is checked
-
-    Raises:
-        ValidationError if extra or read-only fields in payload are found
-    '''
-    # see: https://www.django-rest-framework.org/api-guide/serializers/#object-level-validation
-
-    expected_payload = list(serializer.fields.keys())
-    expected_payload_read_only = [
-        field for field in serializer.fields if serializer.fields[field].read_only
-    ]
-
-    for key in serializer.initial_data.keys():
-        if key not in expected_payload:
-            raise ValidationError(_("Unexpected property in payload"), code=key)
-        if key in expected_payload_read_only:
-            raise ValidationError(_("Found read-only property in payload"), code=key)
